// src2/frontend/providers/platform/ports/types.ts
export type Unsubscribe = () => void
// ...
export interface DebugVariableResult {
  success: boolean
  tick?: number
  lastIndex?: number
  data?: number[]
  error?: string
  needsReconnect?: boolean
}

export interface DebugSetResult {
  success: boolean
  error?: string
}

export interface Md5VerifyResult {
  success: boolean
  match?: boolean
  targetMd5?: string
  error?: string
}